test: deduplicate code
- create a suite_test file to reduce noise in client_test
- create a testRunner structure to drop global variables
- encapsulate common functionality into functions
    - setup code for each test
    - createContainer where we immediately also list to see if the container was created
    - startContainer where we immediately also list to see if the container is running
